+++ CHANGELOG.md
@@ -42,6 +42,7 @@ Breaking changes
 * Shebangs at the very beginning of script files are skipped when loading them.
 * [`smartstring`](https://crates.io/crates/smartstring) is used for identifiers by default. Currently, a PR branch is pulled because it breaks on `no-std` builds. The official crate will be used once `smartstring` is fixed to support `no-std`.
 * `Map` is now an alias to `BTreeMap<SmartString, Dynamic>` instead of `HashMap` because most object maps hold few properties.
+* `EvalAltResult::FnWrongDefinition` is renamed `WrongFnDefinition` for consistency.
 
 New features
 ------------